Extract BaseEnrichmentJob/HttpEnrichmentJob and move enrichment jobs into their own folder
